+++ typedql/schema.ts
@@ -1,3 +1,4 @@
+import { Document, DocumentsInput } from './types/Document'
 import { SchemaResponse } from './types/SchemaResponse'
 
 export interface Query {
@@ -5,4 +6,18 @@ export interface Query {
    * Get schema of CRM
    */
   getSchema: SchemaResponse
+
+}
+
+export interface Mutation {
+  /** @graphql Directives
+   *  @cacheControl(scope: PRIVATE)
+   */
+
+  /**
+   * Save given documents.
+   *
+   * @param documents Documents to be saved.
+   */
+  saveDocuments(documents: DocumentsInput): Document[]
 }