+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add `saveDocuments` mutation.
+
 ## [0.3.0] - 2019-04-25
 
 ### Changed